@@ -18,73 +18,61 @@ classifiers = [
   "License :: OSI Approved :: BSD License",
   "Operating System :: OS Independent",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
-  "Programming Language :: Python :: Implementation :: CPython",
   "Topic :: Scientific/Engineering :: Astronomy",
   "Topic :: Scientific/Engineering :: Physics",
 ]
 # SPEC 0 recommends that packages in the scientific pythoniverse support
 # versions of Python that have been released in the last ≤36 months
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dynamic = ["version"]
 authors = [
   {name = "Joy Velasquez", email = "joy.velasquez@cfa.harvard.edu"},
   {name = "Jonathan Slavin", email = "jslavin@cfa.harvard.edu"},
   {name = "Nick Murpy", email="namurphy@cfa.harvard.edu"},
   {name = "Will Barnes"},
+  {name = "Nabil Freij", email="nabil.freij@gmail.com"},
   {name = "Stuart Mumford"},
 ]
 
 dependencies = [
-  "astropy >= 5.1",
-  "cached-property >= 1.5.2",
+  "astropy >= 5.3.0",
   "matplotlib >= 3.5.0",
-  "numpy >= 1.24.0",
-  "requests >= 2.28.0",
+  "numpy >=1.23.5",
   "scikit-image >= 0.19.0",
-  "scipy >= 1.8.0",
-  # install setuptools to get pkg_resources
-  "setuptools; python_version >= '3.12'",
-  "sunpy[map] >= 4.0.0",
+  # !=1.10.0 due to https://github.com/scipy/scipy/issues/17718
+  "scipy >= 1.9.0 , != 1.10.0",
+  "sunpy[map] >= 5.0.0",
 ]
 
 [project.optional-dependencies]
 dev = [
+  "xrtpy[tests,docs]",
   "nox >= 2022.8.7",
-  "pre-commit >= 3.6.0",
 ]
 tests = [
   "pytest >= 8.0.0",
-  "pytest-allclose >= 1.0.0",
+  "pytest-astropy",
   "pytest-xdist >= 3.6.1",
 ]
 docs = [
-  "docutils >= 0.19",
-  "imageio >= 2.20.0",
-  "ipykernel >= 6.20.0",
-  "ipython >= 8.4.0",
-  "ipywidgets >= 8.1.0",
-  "jinja2 != 3.1, >= 3.0.0",
-  "nbconvert >= 7.7.0, < 7.14",
-  "nbsphinx >= 0.9",
-  "numpydoc >= 1.5.0",
-  "pillow >= 9.1.0",
-  "pygments >= 2.12.0",
+  "pydata_sphinx_theme >= 0.15.0",
+  # Need pkg_resources for sphinxcontrib-bibtex
+  # and this comes from setuptools
+  # This is a bug upstream in the extension which
+  # has yet to be fixed.
+  "setuptools",
   "sphinx >= 7.3.0",
-  "sphinx-changelog >= 1.5.0",
-  "sphinx-codeautolink >= 0.15.2",
+  "sphinx_automodapi >= 0.17.0",
   "sphinx-copybutton >= 0.5.2",
   "sphinx-gallery >= 0.16.0",
   "sphinx-hoverxref >= 1.4.0",
   "sphinx-issues >= 4.1.0",
-  "sphinx_automodapi >= 0.17.0",
-  "sphinx_rtd_theme >= 2.0.0",
   "sphinxcontrib-bibtex >= 2.6.2",
-  "sunpy[net] >= 4.0.0",
-  "towncrier >= 23.11.0",
+  "sphinxext-opengraph>=0.6.0",
+  "sunpy[net] >= 5.0.0",
 ]
 
 [project.urls]
