@@ -1,69 +1,53 @@
 import nox
 
-nox.options.sessions = ["tests", "linters"]
-
-python_versions = ("3.9", "3.10", "3.11", "3.12")
-
-sphinx_paths = ["docs", "docs/_build/html"]
-sphinx_fail_on_warnings = ["-W", "--keep-going"]
-sphinx_builder = ["-b", "html"]
-sphinx_opts = sphinx_paths + sphinx_fail_on_warnings + sphinx_builder
-sphinx_no_notebooks = ["-D", "nbsphinx_execute=never"]
-sphinx_nitpicky = ["-n"]
-
-pytest_options = [
-    "--ignore",
-    "xrtpy/response/effective_area.py",
-    "--ignore",
-    "xrtpy/response/temperature_response.py",
-]
+nox.options.sessions = ["tests"]
+python_versions = ("3.10", "3.11", "3.12")
 
 
-@nox.session(python=python_versions)
+@nox.session
 def tests(session):
-    session.install(".[dev,tests]")
+    """
+    Run tests with pytest.
+    """
+    pytest_options = {}
+    session.install(".[tests]")
     session.run("pytest", *pytest_options)
 
 
 @nox.session
 def linters(session):
+    """
+    Run all pre-commit hooks on all files.
+    """
     session.install("pre-commit")
     session.run("pre-commit", "run", "--all-files", *session.posargs)
 
 
 @nox.session
 def import_package(session):
+    """
+    Import xrtpy.
+    """
     session.install(".")
-    session.run("python", "-c", 'import xrtpy')  # fmt: skip
+    session.run("python", "-c", "import xrtpy")
 
 
 @nox.session
-def build_docs(session):
-    session.install(".[dev,docs]")
-    session.run(
-        "sphinx-build",
-        *sphinx_opts,
-        *session.posargs,
+def docs(session):
+    """
+    Build documentation with Sphinx.
+    """
+    sphinx_paths = ["docs", "docs/_build/html"]
+    sphinx_fail_on_warnings = ["-W", "--keep-going"]
+    sphinx_builder = ["-b", "html"]
+    sphinx_nitpicky = ["-n"]
+    sphinx_opts = (
+        sphinx_paths + sphinx_fail_on_warnings + sphinx_builder + sphinx_nitpicky
     )
-
-
-@nox.session
-def build_docs_nitpicky(session):
-    session.install(".[dev,docs]")
+    session.install(".[docs]")
     session.run(
         "sphinx-build",
         *sphinx_opts,
         *sphinx_nitpicky,
-        *session.posargs,
-    )
-
-
-@nox.session
-def build_docs_no_examples(session):
-    session.install(".[dev,docs]")
-    session.run(
-        "sphinx-build",
-        *sphinx_opts,
-        *sphinx_no_notebooks,
         *session.posargs,
     )
