@@ -10,8 +10,8 @@ from xrtpy import response
 try:
     from xrtpy.version import __version__
 except ImportError:
-    warnings.warn("version not found.")  # noqa: B028
-
+    warnings.warn("version not found.", stacklevel=3)
+    __version__ = "0.0.0"
 
 # Then you can be explicit to control what ends up in the namespace,
-__all__ = ["response"]
+__all__ = ["response", "__version__"]
