@@ -1,8 +1,6 @@
 """Subpackage for image correction"""
 
-__all__ = [
-    "deconvolve",
-    "remove_lightleak",
-]
+from xrtpy.image_correction.deconvolve import deconvolve
+from xrtpy.image_correction.remove_lightleak import remove_lightleak
 
-from xrtpy.image_correction import deconvolve, remove_lightleak
+__all__ = ["deconvolve", "remove_lightleak"]
