@@ -1,6 +1,5 @@
 """Response analysis for Hinode/XRT"""
 
-from xrtpy.response import channel, effective_area, temperature_response
 from xrtpy.response.channel import (
     CCD,
     Channel,
@@ -11,10 +10,10 @@ from xrtpy.response.channel import (
     resolve_filter_name,
 )
 from xrtpy.response.effective_area import EffectiveAreaFundamental
+from xrtpy.response.temperature_from_filter_ratio import temperature_from_filter_ratio
 from xrtpy.response.temperature_response import TemperatureResponseFundamental
 
 __all__ = [
-    "channel",
     "Geometry",
     "EntranceFilter",
     "Mirror",
@@ -23,7 +22,6 @@ __all__ = [
     "Channel",
     "resolve_filter_name",
     "EffectiveAreaFundamental",
-    "effective_area",
-    "temperature_response",
     "TemperatureResponseFundamental",
+    "temperature_from_filter_ratio",
 ]
