@@ -1,6 +1,7 @@
 from datetime import datetime
 from pathlib import Path
 
+import numpy as np
 import pytest
 from astropy import units as u
 
@@ -70,9 +71,7 @@ def test_EffectiveArea_filter_name(name):
     instance = EffectiveAreaFundamental(
         name, datetime(year=2013, month=9, day=22, hour=22, minute=0, second=0)
     )
-    actual_attr_value = instance.name
-
-    assert actual_attr_value == name
+    assert instance.name == name
 
 
 @pytest.mark.parametrize("date", valid_dates)
@@ -92,81 +91,59 @@ def test_EffectiveArea_contamination_on_filter(name, date):
 @pytest.mark.parametrize("date", invalid_dates)
 @pytest.mark.parametrize("name", channel_names)
 def test_EffectiveArea_exception_is_raised(name, date):
-    with pytest.raises(ValueError):  # noqa: PT011
+    with pytest.raises(ValueError, match="Invalid date"):
         EffectiveAreaFundamental(name, date)
 
 
-def get_IDL_data_files():
-    directory = (
-        Path(__file__).parent.parent.absolute()
-        / "data"
-        / "effective_area_IDL_testing_files"
-    )
-    filter_data_files = directory.glob("**/*.txt")
+"""def get_IDL_data_files():
+    filter_data_files = []
+    for dir in get_pkg_data_filenames(
+        "data/effective_area_IDL_testing_files", package="xrtpy.response.tests"
+    ):
+        filter_data_files += list(Path(dir).glob("*.txt"))
     return sorted(filter_data_files)
+"""
 
 
-filenames = get_IDL_data_files()
+def get_IDL_data_files():
+    data_dir = Path(__file__).parent / "data" / "effective_area_IDL_testing_files"
+    assert data_dir.exists(), f"Data directory {data_dir} does not exist."
+    files = sorted(data_dir.glob("**/*.txt"))
+    # print(f"\n\n\nFound files: {files}\n\n")  # Debugging output
+    return files
 
 
-def _IDL_raw_data_list(filename):
-    with open(filename) as filter_file:  # noqa: PTH123
-        list_of_IDL_effective_area_data = []
-        for line in filter_file:
-            stripped_line = line.strip()
-            line_list = stripped_line.split()
-            list_of_IDL_effective_area_data.append(line_list)
+# NOTE: This is marked as xfail because the IDL results that this test compares against
+# are incorrect due to the use of quadratic interpolation in the contamination curves
+# which leads to ringing near the edges in the contamination curve.
+# See https://github.com/HinodeXRT/xrtpy/pull/284#issuecomment-2334503108
+# @pytest.mark.xfail
+@pytest.mark.parametrize("filename", get_IDL_data_files())
+def test_effective_area_compare_idl(filename):
+    with filename.open() as f:
+        filter_name = f.readline().split()[1]
+        filter_obs_date = " ".join(f.readline().split()[1:])
+    # NOTE: Annoyingly the date strings use "Sept" instead of "Sep" for "September"
+    filter_obs_date = filter_obs_date.replace("Sept", "Sep")
+    IDL_data = np.loadtxt(filename, skiprows=3)
+    IDL_wavelength = IDL_data[:, 0] * u.AA
+    IDL_effective_area = IDL_data[:, 1] * u.cm**2
 
-    return list_of_IDL_effective_area_data
-
-
-def IDL_test_filter_name(list_of_lists):
-    return str(list_of_lists[0][1])
-
-
-def IDL_test_date(list_of_lists):
-    obs_date = str(list_of_lists[1][1])
-    obs_time = str(list_of_lists[1][2])
-
-    day = int(obs_date[:2])
-
-    month_datetime_object = datetime.strptime(obs_date[3:6], "%b")
-    month = month_datetime_object.month
-
-    year = int(obs_date[8:12])
-
-    hour = int(obs_time[:2])
-    minute = int(obs_time[3:5])
-    second = int(obs_time[6:8])
-
-    return datetime(year, month, day, hour, minute, second)
-
-
-def _IDL_effective_area_raw_data(filename):
-    with open(filename) as filter_file:  # noqa: PTH123
-        list_of_lists = []
-        for line in filter_file:
-            stripped_line = line.strip()
-            line_list = stripped_line.split()
-            list_of_lists.append(line_list)
-
-    effective_area = [list_of_lists[i][1] for i in range(3, len(list_of_lists))]
-    effective_area = [float(i) for i in effective_area] * u.cm**2
-
-    return effective_area
-
-
-@pytest.mark.parametrize("filename", filenames)
-def test_EffectiveAreaPreparatory_effective_area(filename, allclose):
-    data_list = _IDL_raw_data_list(filename)
-
-    filter_name = IDL_test_filter_name(data_list)
-    filter_obs_date = IDL_test_date(data_list)
-
-    IDL_effective_area = _IDL_effective_area_raw_data(filename)
-
+    # Interpolate XRTpy effective area onto the IDL wavelength grid
     instance = EffectiveAreaFundamental(filter_name, filter_obs_date)
     actual_effective_area = instance.effective_area()
+    XRTpy_effective_area = (
+        np.interp(
+            IDL_wavelength.value,  # Target grid (IDL wavelengths)
+            instance.wavelength.value,  # Source grid (XRTpy wavelengths)
+            actual_effective_area.value,  # Data to interpolate
+        )
+        * u.cm**2
+    )
 
-    assert actual_effective_area.unit == IDL_effective_area.unit
-    assert allclose(actual_effective_area.value, IDL_effective_area.value, atol=1e-2)
+    assert u.allclose(
+        XRTpy_effective_area,  # Interpolated XRTpy values
+        IDL_effective_area,  # Original IDL values
+        rtol=1e-4,  # Relative tolerance
+        atol=1.0e-4 * u.cm**2,
+    )
