@@ -1,8 +1,6 @@
-__all__ = [
-    "epoch",
-]
-
 import astropy.time
+
+__all__ = ["epoch"]
 
 # Hinode-XRT mission elapsed time "Epoch" is Sept 22, 2006 21:36:00.
 epoch = astropy.time.Time("2006-09-22 21:36:00")
